Move test data under data/basic/
