+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/JsonFileToTabularFileConverterTest.kt
@@ -1,21 +1,16 @@
 package cz.dynawest.csvcruncher.test.json
 
-import cz.dynawest.csvcruncher.converters.FlattenedEntrySequence
-import cz.dynawest.csvcruncher.converters.EntryProcessor
-import cz.dynawest.csvcruncher.converters.JsonFileToTabularFileConverter
 import cz.dynawest.csvcruncher.converters.MyProperty
+import cz.dynawest.csvcruncher.test.json.JsonTestUtils.prepareEntriesFromFile
 import cz.dynawest.csvcruncher.util.logger
-import cz.dynawest.util.ResourceLoader
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
-import java.io.InputStream
-import java.nio.file.Path
 
 class JsonFileToTabularFileConverterTest {
 
     @Test
     fun testConvertJson_01_arrayAtRoot_sameProperties() {
-        val entries = prepareEntriesFromFile("01-arrayAtRoot-sameProperties.json", "/")
+        val entries = prepareEntriesFromFile("data/basic/01-arrayAtRoot-sameProperties.json", "/")
 
         Assertions.assertThat(entries).size().isEqualTo(3)
         Assertions.assertThat(entries).element(0).extracting({it["id"]}).isEqualTo(MyProperty.String("id", "1"))
@@ -26,7 +21,7 @@ class JsonFileToTabularFileConverterTest {
 
     @Test
     fun `testConvertJson_02-arrayAtRoot-propertiesMissing`() {
-        val entries = prepareEntriesFromFile("02-arrayAtRoot-propertiesMissing.json", "/")
+        val entries = prepareEntriesFromFile("data/basic/02-arrayAtRoot-propertiesMissing.json", "/")
 
         Assertions.assertThat(entries).size().isEqualTo(3)
         Assertions.assertThat(entries).element(0).extracting({it["id"]}).isEqualTo(MyProperty.String("id", "1"))
@@ -39,7 +34,7 @@ class JsonFileToTabularFileConverterTest {
 
     @Test
     fun `testConvertJson_03-arrayAtRoot-varyingProperties`() {
-        val entries = prepareEntriesFromFile("03-arrayAtRoot-varyingProperties.json", "/")
+        val entries = prepareEntriesFromFile("data/basic/03-arrayAtRoot-varyingProperties.json", "/")
 
         Assertions.assertThat(entries).size().isEqualTo(3)
         Assertions.assertThat(entries).element(0).extracting({it["id"]}).isEqualTo(MyProperty.String("id", "1"))
@@ -55,7 +50,7 @@ class JsonFileToTabularFileConverterTest {
 
     @Test
     fun `testConvertJson_04-arrayAtRoot-differentTypes`() {
-        val entries = prepareEntriesFromFile("04-arrayAtRoot-differentTypes.json", "/")
+        val entries = prepareEntriesFromFile("data/basic/04-arrayAtRoot-differentTypes.json", "/")
 
         Assertions.assertThat(entries).size().isEqualTo(5)
         Assertions.assertThat(entries).element(0).extracting({it["id"]}).isEqualTo(MyProperty.String("id", "1"))
@@ -68,7 +63,7 @@ class JsonFileToTabularFileConverterTest {
 
     @Test
     fun `testConvertJson_10-arrayAtRoot-nestedObject`() {
-        val entries = prepareEntriesFromFile("10-arrayAtRoot-nestedObject.json", "/")
+        val entries = prepareEntriesFromFile("data/basic/10-arrayAtRoot-nestedObject.json", "/")
 
         Assertions.assertThat(entries).size().isEqualTo(3)
         Assertions.assertThat(entries).element(0).extracting({it["id"]}).isEqualTo(MyProperty.String("id", "1"))
@@ -84,7 +79,7 @@ class JsonFileToTabularFileConverterTest {
 
     @Test
     fun `testConvertJson_11-arrayAtRoot-nestedArray`() {
-        val entries = prepareEntriesFromFile("11-arrayAtRoot-nestedArray.json", "/")
+        val entries = prepareEntriesFromFile("data/basic/11-arrayAtRoot-nestedArray.json", "/")
 
         Assertions.assertThat(entries).size().isEqualTo(3)
         Assertions.assertThat(entries).element(0).extracting({it["id"]}).isEqualTo(MyProperty.String("id", "1"))
@@ -99,32 +94,13 @@ class JsonFileToTabularFileConverterTest {
 
     @Test
     fun `testConvertJson_12-arrayInObject-sameProperties`() {
-        val entries = prepareEntriesFromFile("12-arrayInObject-sameProperties.json", "/items")
+        val entries = prepareEntriesFromFile("data/basic/12-arrayInObject-sameProperties.json", "/items")
 
         Assertions.assertThat(entries).size().isEqualTo(3)
         Assertions.assertThat(entries).element(0).extracting({it["id"]}).isEqualTo(MyProperty.String("id", "1"))
         Assertions.assertThat(entries).element(0).extracting({it["name"]}).isEqualTo(MyProperty.String("name", "Ondra"))
         Assertions.assertThat(entries).element(1).extracting({it["id"]}).isEqualTo(MyProperty.String("id", "2"))
         Assertions.assertThat(entries).element(2).extracting({it["id"]}).isEqualTo(MyProperty.String("id", "3"))
-    }
-
-
-
-
-
-    private fun prepareEntriesFromFile(testFilePath: String, itemsArraySprout: String = "/"): MutableList<Map<String, MyProperty>> {
-        val converter = JsonFileToTabularFileConverter()
-        val inputStream: InputStream = ResourceLoader.openResourceAtRelativePath(Path.of(testFilePath))
-
-        val entries = mutableListOf<Map<String, MyProperty>>()
-        converter.processEntries(inputStream, Path.of(itemsArraySprout), object : EntryProcessor {
-            override fun collectPropertiesMetadata(entry: FlattenedEntrySequence) {
-                val entryMap = entry.flattenedProperties.associateBy { myProp -> myProp.name }
-                entries.add(entryMap)
-                log.info("Entry: ${entryMap}")
-            }
-        })
-        return entries
     }
 
     companion object { private val log = logger() }
