+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/data/basic/01-arrayAtRoot-sameProperties.json
