+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/data/basic/02-arrayAtRoot-propertiesMissing.json
