+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/data/basic/03-arrayAtRoot-varyingProperties.json
