+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/data/basic/04-arrayAtRoot-differentTypes.json
