+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/data/basic/10-arrayAtRoot-nestedObject.json
