+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/data/basic/11-arrayAtRoot-nestedArray.json
