+++ src/test/kotlin/cz/dynawest/csvcruncher/test/json/data/basic/12-arrayInObject-sameProperties.json
